some ISTL compliance

+++ stuff/functions.hh
@@ -675,9 +675,9 @@ void divideFunctionDofs( DiscreteFunctionType& target, const DiscreteFunctionTyp
 template < class MatrixType >
 bool MatrixContainsNanOrInf( const MatrixType& matrix )
 {
-    for(int row=0; row < int(matrix.rows()); ++row) {
-        for(int col=0; col < int(matrix.cols()); ++col) {
-            if ( std::isnan( matrix(row,col) ) || std::isinf( matrix(row,col) ) )
+    for(size_t row=0; row < matrix.rows(); ++row) {
+        for(size_t col=0; col < matrix.cols(); ++col) {
+			if ( std::isnan( matrix(row,col) ) || std::isinf( matrix(row,col) ) )
 				return true;
 		}
 	}
